domain

+++ data/src/main/java/com/tailspin/data/source/local/entity/ItemEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "item")
 data class ItemEntity (
-    @PrimaryKey var id : Long,
+    @PrimaryKey var id : Int,
     var name : String,
     var time : Int,
     var amount : Int
